@@ -1,44 +1,59 @@
 name: spot
 description: A new Flutter project.
-
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
-
+publish_to: "none"
 version: 1.1.0+1
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: ">=2.17.3 <3.0.0"
 
 dependencies:
   cupertino_icons: ^1.0.2
-  dio: ^4.0.4
+  dio: ^4.0.6
   flutter:
     sdk: flutter
-  freezed_annotation: ^1.1.0
+  freezed_annotation: ^2.1.0
   hive_flutter: ^1.1.0
   intl: ^0.17.0
-  json_annotation: ^4.4.0
+  json_annotation: ^4.6.0
   logger: ^1.1.0
   responsive_builder: ^0.4.2
-  rive: ^0.8.4
+  rive: ^0.9.1
   shimmer: ^2.0.0
-  stacked: ^2.2.8
+  stacked: ^2.3.15
   stacked_hooks: ^0.2.1+1
-  stacked_services: 0.8.22+1
-  stacked_themes: ^0.3.8+1
-  url_launcher: ^6.0.20
-
-dependency_overrides:
-  get: 4.6.1
+  stacked_services: ^0.9.3
+  stacked_themes: ^0.3.9
+  url_launcher: ^6.1.5
 
 dev_dependencies:
-  build_runner: ^2.1.7
-  flutter_lints: ^1.0.4
+  build_runner: ^2.2.0
+  flutter_launcher_icons: ^0.11.0
+  flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
-  freezed: ^1.1.1
-  hive_generator: ^1.1.2
-  json_serializable: ^6.1.4
-  stacked_generator: ^0.5.7
+  freezed: ^2.1.0+1
+  hive_generator: ^1.1.3
+  json_serializable: ^6.3.1
+  stacked_generator: ^0.7.13
+
+flutter_icons:
+  android: "ic_launcher"
+  ios: true
+  remove_alpha_ios: true
+  image_path: "assets/logo/spot.png"
+  min_sdk_android: 21 # android min sdk min:16, default 21
+  web:
+    generate: true
+    image_path: "assets/logo/spot.png"
+    background_color: "#22242C"
+    theme_color: "#22242C"
+  windows:
+    generate: true
+    image_path: "assets/logo/spot.png"
+    icon_size: 48 # min:48, max:256, default: 48
+  macos:
+    generate: true
+    image_path: "assets/logo/spot.png"
 
 flutter:
   uses-material-design: true
