@@ -4,11 +4,11 @@
 // StackedRouterGenerator
 // **************************************************************************
 
-// ignore_for_file: public_member_api_docs
+// ignore_for_file: public_member_api_docs, unused_import, non_constant_identifier_names
 
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
-import 'package:stacked/stacked_annotations.dart';
+import 'package:stacked_services/stacked_services.dart';
 
 import '../ui/views/about/about_view.dart';
 import '../ui/views/home/home_view.dart';
@@ -38,24 +38,87 @@ class StackedRouter extends RouterBase {
   final _pagesMap = <Type, StackedRouteFactory>{
     HomeView: (data) {
       return PageRouteBuilder<dynamic>(
-        pageBuilder: (context, animation, secondaryAnimation) =>
-            const HomeView(),
-        settings: data,
-      );
+          pageBuilder: (context, animation, secondaryAnimation) =>
+              const HomeView(),
+          settings: data,
+          transitionsBuilder: data.transition ??
+              (context, animation, secondaryAnimation, child) {
+                return child;
+              });
     },
     RecentsView: (data) {
       return PageRouteBuilder<dynamic>(
-        pageBuilder: (context, animation, secondaryAnimation) =>
-            const RecentsView(),
-        settings: data,
-      );
+          pageBuilder: (context, animation, secondaryAnimation) =>
+              const RecentsView(),
+          settings: data,
+          transitionsBuilder: data.transition ??
+              (context, animation, secondaryAnimation, child) {
+                return child;
+              });
     },
     AboutView: (data) {
       return PageRouteBuilder<dynamic>(
-        pageBuilder: (context, animation, secondaryAnimation) =>
-            const AboutView(),
-        settings: data,
-      );
+          pageBuilder: (context, animation, secondaryAnimation) =>
+              const AboutView(),
+          settings: data,
+          transitionsBuilder: data.transition ??
+              (context, animation, secondaryAnimation, child) {
+                return child;
+              });
     },
   };
+}
+
+/// ************************************************************************
+/// Extension for strongly typed navigation
+/// *************************************************************************
+
+extension NavigatorStateExtension on NavigationService {
+  Future<dynamic> navigateToHomeView({
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  }) async {
+    return navigateTo(
+      Routes.homeView,
+      id: routerId,
+      preventDuplicates: preventDuplicates,
+      parameters: parameters,
+      transition: transition,
+    );
+  }
+
+  Future<dynamic> navigateToRecentsView({
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  }) async {
+    return navigateTo(
+      Routes.recentsView,
+      id: routerId,
+      preventDuplicates: preventDuplicates,
+      parameters: parameters,
+      transition: transition,
+    );
+  }
+
+  Future<dynamic> navigateToAboutView({
+    int? routerId,
+    bool preventDuplicates = true,
+    Map<String, String>? parameters,
+    Widget Function(BuildContext, Animation<double>, Animation<double>, Widget)?
+        transition,
+  }) async {
+    return navigateTo(
+      Routes.aboutView,
+      id: routerId,
+      preventDuplicates: preventDuplicates,
+      parameters: parameters,
+      transition: transition,
+    );
+  }
 }
