@@ -66,7 +66,7 @@ class TimeHelper {
       result = _countYears(difference);
     }
 
-    return !result.startsWith('J') ? result + ' ago' : result;
+    return !result.startsWith('J') ? '$result ago' : result;
   }
 
   /// Converts the time difference to a number of seconds.
@@ -74,7 +74,7 @@ class TimeHelper {
   ///   returns ("Just now" OR "X seconds")
   static String _countSeconds(int difference) {
     int count = (difference / 1000).truncate();
-    return count > 1 ? count.toString() + ' seconds' : 'Just now';
+    return count > 1 ? '$count seconds' : 'Just now';
   }
 
   /// Converts the time difference to a number of minutes.
